Fixed compiler warning.
git-svn-id: https://svn.opendtect.org/od/trunk@27337 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Seis/segyhdr.h
@@ -96,7 +96,7 @@ public:
     inline void	setEntryVal( int idx, int val )
 		{ return hdrDef()[idx]->putValue(buf_,val); }
 
-    short	format() const		{ return entryVal(EntryFmt()); }
+    short	format() const		{ return (short) entryVal(EntryFmt()); }
     int		nrSamples() const	{ return entryVal(EntryNs()); }
     float	sampleRate(bool isdpth) const;
     bool	isInFeet() const	{ return entryVal(EntryMFeet()) == 2; }
